requirements.txt
ccxt
pandas
numpy
python-dotenv
requests
pyyaml
